moves schemas into one place
