+++ openvair/modules/backup/schemas.py
@@ -10,6 +10,12 @@ from typing import List, Optional
 
 from pydantic import BaseModel
 
+from openvair.modules.backup.config import (
+    RESTIC_DIR,
+    STORAGE_DATA,
+    RESTIC_PASSWORD,
+)
+
 
 class ResticBackupResult(BaseModel):
     """Schema for restic backup results.
@@ -113,3 +119,38 @@ class ResticSnapshot(BaseModel):
         """
 
         extra = 'ignore'
+
+
+class ResticBackuperData(BaseModel):
+    """Schema for Restic backuper configuration data.
+
+    This schema defines default paths and passwords required to configure
+    a Restic backuper.
+
+    Attributes:
+        source_path (str): Path to the source data to be backed up. Defaults
+            to the `STORAGE_DATA` configuration value.
+        restic_path (str): Path to the Restic repository. Defaults to the
+            `RESTIC_DIR` configuration value.
+        restic_password (str): Password for the Restic repository. Defaults
+            to the `RESTIC_PASSWORD` configuration value.
+    """
+
+    source_path: str = str(STORAGE_DATA)
+    restic_path: str = str(RESTIC_DIR)
+    restic_password: str = str(RESTIC_PASSWORD)
+
+
+class DataForResticManager(ResticBackuperData):
+    """Schema for preparing data for the Restic domain manager.
+
+    This schema wraps Restic backuper configuration data along with the
+    backuper type, used to identify and manage Restic-related operations.
+
+    Attributes:
+        backuper_type (str): Type of the backuper. Defaults to "restic".
+        backuper_data (Dict): Configuration data for the Restic backuper,
+            generated from the `ResticBackuperData` schema.
+    """
+
+    backuper_type: str = 'restic'
